fix(css_parser): fix Constant crashes when editing css files #3256
