+++ crates/biome_css_parser/tests/css_test_suite/error/property/color_error.css
@@ -1,0 +1,3 @@
+.formTable tbody td {
+	border-left: 1px # solid;
+}
