+++ crates/biome_css_parser/tests/spec_test.rs
@@ -174,8 +174,9 @@ pub fn run(test_case: &str, _snapshot_name: &str, test_directory: &str, outcome_
 #[test]
 pub fn quick_test() {
     let code = r#"
-@font-face { color: U+0100-024F; }
-
+.formTable tbody td {
+    border-left: 1px # solid;
+}
     "#;
 
     let root = parse_css(
